Move some functions to formatting package
